requirements.txt
# Social Media APIs
tweepy==4.14.0
praw==7.7.1
requests==2.31.0

# Data Processing
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1

# NLP and Sentiment Analysis
transformers==4.37.2
torch==2.2.1
nltk==3.8.1
spacy==3.7.4
emoji==2.10.1

# Machine Learning
scikit-learn==1.6.1
xgboost==2.0.3
prophet==1.1.5
tensorflow==2.16.1

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
streamlit==1.32.0

# Other
pytest==8.0.2 